pyproject.toml
[project]
name = "bard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.31.0",
    "email-validator>=2.1.0",
    "fastapi[standard]>=0.128.0",
    "passlib[argon2]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "langchain>=1.2.0",
    "langchain-openai>=0.3.0",
    "pillow>=10.2.0",
    "pillow-heif>=0.15.0",
    "httpx>=0.26.0",
    "numpy>=2.4.1",
    "opencv-python>=4.11.0.86",
    "deepface>=0.0.96",
    "retina-face>=0.0.17",
    "supabase>=2.27.1",
    "tf-keras>=2.20.1",
    "pgvector>=0.4.2",
    "celery>=5.6.2",
    "redis>=7.1.0",
]
